Move old GalleryItem
