+++ src/components/images/ImageLayoutGrid.tsx
@@ -6,7 +6,7 @@ import {type AppBskyEmbedImages} from '@atproto/api'
 import {type Dimensions} from '#/view/com/lightbox/ImageViewing/@types'
 import {atoms as a, useBreakpoints} from '#/alf'
 import {PostEmbedViewContext} from '#/components/Post/Embed/types'
-import {GalleryItem} from './Gallery/GalleryItem'
+import {GalleryItem} from './ImageLayoutGridItem'
 
 interface ImageLayoutGridProps {
   images: AppBskyEmbedImages.ViewImage[]
